Done Some Score Stuff
Finished adding in the zones for detecting where the player has dropped the nuts and bolts and using this the score system checks if they've done a correct or incorrect sort or if they've missed the chance to sort.

Based on this a live is taken away or score given.